% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMun.R
\name{amostraBrasil}
\alias{amostraBrasil}
\title{Generates Brazil's IBGE Householding Sample}
\usage{
amostraBrasil(codibge = "", municipio = "", listasetor = c(), N = 0,
  geocod = F, shape = F, saveall = T, dados = NULL)
}
\arguments{
\item{codibge}{A string}

\item{municipio}{A string}

\item{listasetor}{A string vector}

\item{N}{A number}

\item{geocod}{A boolean}

\item{shape}{A boolean}

\item{saveall}{A boolean}

\item{dados}{A null}
}
\value{
A dataframe with N addresses sampled from IBGE census
}
\description{
Generates Brazil's IBGE Householding Sample
}
\examples{
amostraBrasil(municipio="Pindoba", N=5, geocod=TRUE, shape=TRUE)

}

