% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirFTP.R
\name{dirFTP}
\alias{dirFTP}
\title{List Files in FTP Site}
\usage{
dirFTP(myURL = "", codibge = "")
}
\arguments{
\item{myURL}{A string}

\item{codibge}{A string}
}
\value{
A dataframe with IBGE files for one municipality
}
\description{
List Files in FTP Site
}

