% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeoCodeCS.R
\name{getGeoCodeCS}
\alias{getGeoCodeCS}
\title{Geodode an address by Google Maps}
\usage{
getGeoCodeCS(gcStr)
}
\arguments{
\item{gcStr}{A string}
}
\value{
Geocoded address
}
\description{
Geodode an address by Google Maps
}

