% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIBGEMunSHP.R
\name{getIBGEMunSHP}
\alias{getIBGEMunSHP}
\title{Get a Shapefile From IBGE Website}
\usage{
getIBGEMunSHP(codibge = "", nomemun = "")
}
\arguments{
\item{codibge}{A string}

\item{nomemun}{A string}
}
\value{
Shapefile of Municipality in Brazil
}
\description{
Get a Shapefile From IBGE Website
}

