% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUNICIPIOS.R
\docType{data}
\name{MUNICIPIOS.IBGE}
\alias{MUNICIPIOS.IBGE}
\title{Names and IBGE Codes of All Brazilian municipalities.}
\format{A data frame}
\source{
\url{http://www.ibge.gov.br/}
}
\usage{
MUNICIPIOS.IBGE
}
\description{
A dataset containing the names and IBGE codes of all municipalities in Brazil
}
\details{
\itemize{
  \item UF. State
  \item MUNICIPIO. Municipality name
  \item CODIBGE. Municipality IBGE code
}
}
\keyword{datasets}

