% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDentroFora.R
\name{setDentroFora}
\alias{setDentroFora}
\title{Check if Points as Inside a Polygon}
\usage{
setDentroFora(pontos, limites)
}
\arguments{
\item{pontos}{A vector}

\item{limites}{A vector}
}
\value{
Vector of points
}
\description{
Check if Points as Inside a Polygon
}

