% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLatLong.R
\name{setLatLong}
\alias{setLatLong}
\title{Geocode addresses by Google Maps}
\usage{
setLatLong(datafr, addresscol)
}
\arguments{
\item{datafr}{A dataframe}

\item{addresscol}{A vector}
}
\value{
The submited dataframe plus columns Lat, Lng, Address (Google Maps style) and Status
}
\description{
Geocode addresses by Google Maps
}

